chore: Move solidity-specific unit-test for IPA into separate module
