+++ src/provider/mod.rs
@@ -19,6 +19,7 @@ pub(crate) mod util;
 
 // crate-private modules
 mod keccak;
+mod tests;
 
 use crate::{
   provider::{
@@ -167,7 +168,7 @@ impl CurveCycleEquipped for PallasEngine {
 }
 
 #[cfg(test)]
-mod tests {
+mod test {
   use crate::provider::{
     bn256_grumpkin::{bn256, grumpkin},
     secp_secq::{secp256k1, secq256k1},
